+++ internal/sqlutil/doc.go
@@ -1,0 +1,2 @@
+// Package sqlutil provides utilities to SQL-related instrumentation modules.
+package sqlutil