+++ module/apmgorm/doc.go
@@ -1,0 +1,2 @@
+// Package apmgorm provides wrappers for tracing GORM operations.
+package apmgorm